requirements.txt
# Core dependencies
requests>=2.31.0

# Configuration management
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
